+++ src/main/java/com/smart/parking/repository/CarRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
 
     @Query(nativeQuery = true, value = "SELECT * FROM car WHERE number_plate =:licence")
     Optional<Car> findByLicense(@Param("licence") String license);
 
     @Query(nativeQuery = true, value = "SELECT * FROM car WHERE user_id =:user_id")
-    Optional<Car> findByUserId(@Param("user_id") Integer userId);
+    Optional<Car> findByUserId(@Param("user_id") Long userId);
 }
